+++ crates/core/src/lib.rs
@@ -2,12 +2,14 @@
 
 mod document;
 mod pattern;
+mod studies;
 mod theory;
 mod time;
 
 pub mod prelude {
     pub use super::document::*;
     pub use super::pattern::*;
+    pub use super::studies::*;
     pub use super::theory::*;
     pub use super::time::*;
 }